Update NOTE - client contact (call center).vbs
Added check box for user to select if they reminded client about the importance of completing the CAF Part I to set the application date.  This has been cited in many FSME reviews during the DHS "fake" call as not being done.  Good for support staff and workers as a reminder.  If checked, will add this to case note.