NOTES - CHANGE REPORT FORM RECEIVED.vbs
This is a script that Financial Workers can use to enter information they have received on a Change Report Form DHS 2402.  The script will case note info entered and also has the option to navigate to DAIL/WRIT to create a new blank TIKL message, automatically sets the TIKL date out 10 days from the date the note is created.