Create Notes - IEVS NOTICE.vbs
This script creates a case note for IEVS matches. There is a separate dialog box if the IEVS match is resolved and if the code is non cooperation then there is also a message box that pops up at the end to remind the user to add a DISQ  panel.