Update NOTES - CHANGE REPORT FORM RECEIVED.vbs
@TheVKC, @IlseFerris
What do you guys think of this? 
This script changes the file from Change Report Form Received to Change Reported. Comes included with combo box to choose how the change was reported (CRF, Phone contact, Email, Etc). The header of the case note will show what type of change was reported based on what fields the worker entered. Also added was a checkbox next to each field to case note if we have verification for the change.