Adding functionality to read interview
BLIP: Added functionality to check interview status off of PROG. It will make a new column with a Y or N if the client has been interviewed for SNAP. 

Functionality for the checking what the results of the expedited screening is included as well but commented out due to concerns over validity of data. Both have been tested and do work. However more than half of the screenings were shown to have the opposite expedited status in small scale double checking.