Resolves #2060 - Adding Verification Still Needed
BLIP: This script is meant to be used when several verifications have been received from the client but there is still outstanding information needed to determine eligibility. It will case note all outstanding information entered in to the script and then open a Word document, and list the information there as well to be sent out to the client; informing them that more information is still needed.